+++ src/main/java/com/example/todoapp/service/post/PostEntity.java
@@ -1,7 +1,7 @@
 package com.example.todoapp.service.post;
 
 public record PostEntity(
-        long user_id,
+        int user_id,
         String content,
         String status,
         String created_at,
